requirements.txt
pandas>=2.0.0
numpy>=1.24.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
